# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers/profile": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get customer profile",
                "description": "Get customer profile",
                "operationId": "d3c979277b503b4cc2c14ee9d13d47c6",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customers/auth/social-register": {
            "post": {
                "tags": [
                    "Customers Social Authentication"
                ],
                "summary": "Customer social register",
                "description": "Customer social register",
                "operationId": "35e81d6867f506640cb0a57a1534256b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "avatar.jpg"
                                    },
                                    "provider": {
                                        "type": "string",
                                        "enum": [
                                            "google",
                                            "facebook",
                                            "twitter"
                                        ],
                                        "example": "google"
                                    },
                                    "provider_id": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/customers/auth/social-login": {
            "post": {
                "tags": [
                    "Customers Social Authentication"
                ],
                "summary": "Customer social login",
                "description": "Customer social login",
                "operationId": "d5d53b53af4d27577cb34a723d6d6c98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "provider": {
                                        "type": "string",
                                        "enum": [
                                            "google",
                                            "facebook",
                                            "twitter"
                                        ],
                                        "example": "google"
                                    },
                                    "provider_id": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}